move account header components to main components directory
