+++ app/javascript/mastodon/components/account_header/memorial_note.tsx
