+++ app/javascript/mastodon/components/account_header/styles.module.scss
