+++ app/javascript/mastodon/features/account_featured/index.tsx
@@ -6,26 +6,26 @@ import { useHistory } from 'react-router';
 
 import { List as ImmutableList } from 'immutable';
 
-import AddIcon from '@/material-icons/400-24px/add.svg?react';
-import { fetchEndorsedAccounts } from 'mastodon/actions/accounts';
-import { AccountListItem } from 'mastodon/components/account_list_item';
-import { ColumnBackButton } from 'mastodon/components/column_back_button';
-import { LoadingIndicator } from 'mastodon/components/loading_indicator';
-import { RemoteHint } from 'mastodon/components/remote_hint';
+import { fetchEndorsedAccounts } from '@/mastodon/actions/accounts';
+import { AccountHeader } from '@/mastodon/components/account_header';
+import { AccountListItem } from '@/mastodon/components/account_list_item';
+import { ColumnBackButton } from '@/mastodon/components/column_back_button';
+import { LoadingIndicator } from '@/mastodon/components/loading_indicator';
+import { RemoteHint } from '@/mastodon/components/remote_hint';
 import {
   Article,
   ItemList,
   Scrollable,
-} from 'mastodon/components/scrollable_list/components';
-import type { TruncatedListItemInfo } from 'mastodon/components/truncated_list';
-import { TruncatedListItems } from 'mastodon/components/truncated_list';
-import { AccountHeader } from 'mastodon/features/account_timeline/components/account_header';
-import BundleColumnError from 'mastodon/features/ui/components/bundle_column_error';
-import Column from 'mastodon/features/ui/components/column';
-import { useAccount } from 'mastodon/hooks/useAccount';
-import { useAccountId } from 'mastodon/hooks/useAccountId';
-import { useAccountVisibility } from 'mastodon/hooks/useAccountVisibility';
-import { useAppDispatch, useAppSelector } from 'mastodon/store';
+} from '@/mastodon/components/scrollable_list/components';
+import type { TruncatedListItemInfo } from '@/mastodon/components/truncated_list';
+import { TruncatedListItems } from '@/mastodon/components/truncated_list';
+import BundleColumnError from '@/mastodon/features/ui/components/bundle_column_error';
+import Column from '@/mastodon/features/ui/components/column';
+import { useAccount } from '@/mastodon/hooks/useAccount';
+import { useAccountId } from '@/mastodon/hooks/useAccountId';
+import { useAccountVisibility } from '@/mastodon/hooks/useAccountVisibility';
+import { useAppDispatch, useAppSelector } from '@/mastodon/store';
+import AddIcon from '@/material-icons/400-24px/add.svg?react';
 
 import { useAccountCollections } from '../collections';
 import { CollectionListItem } from '../collections/components/collection_list_item';
