+++ app/javascript/mastodon/features/account_timeline/v2/index.tsx
@@ -12,8 +12,10 @@ import {
   expandTimelineByKey,
   timelineKey,
 } from '@/mastodon/actions/timelines_typed';
+import { AccountHeader } from '@/mastodon/components/account_header';
 import { Column } from '@/mastodon/components/column';
 import { ColumnBackButton } from '@/mastodon/components/column_back_button';
+import { LimitedAccountHint } from '@/mastodon/components/limited_account_hint';
 import { LoadingIndicator } from '@/mastodon/components/loading_indicator';
 import { RemoteHint } from '@/mastodon/components/remote_hint';
 import StatusList from '@/mastodon/components/status_list';
@@ -25,9 +27,6 @@ import {
 import { useAccountVisibility } from '@/mastodon/hooks/useAccountVisibility';
 import { selectTimelineByKey } from '@/mastodon/selectors/timelines';
 import { useAppDispatch, useAppSelector } from '@/mastodon/store';
-
-import { AccountHeader } from '../components/account_header';
-import { LimitedAccountHint } from '../components/limited_account_hint';
 
 import { AccountTimelineProvider, useAccountContext } from './context';
 import { FeaturedTags } from './featured_tags';
